+++ pytorch_lightning/core/memory.py
@@ -208,7 +208,6 @@ class ModelSummary(object):
 
         input_ = model.example_input_array
         input_ = model.transfer_batch_to_device(input_, model.device)
-        input_ = apply_to_collection(input_, torch.Tensor, lambda x: x.type(model.dtype))
 
         if trainer is not None and trainer.use_amp:
             if NATIVE_AMP_AVALAIBLE: